requirements.txt
APScheduler
requests
python-dotenv